+++ src/Websocket.hpp
@@ -29,7 +29,7 @@ public:
 	void clear();
 	void print(const uint8_t& col, const uint8_t& row, const std::string& s);
 	void flush();
-	void sendAudio(float* audioBuffer, size_t len);
+	void sendAudio(int16_t* audioBuffer, size_t len);
 	void updateParameter(const string& name, const float& value);
 	bool isAudioStreamEnabled() {
 		return audioStreamEnabled_;
